Bump plugin and dependency versions
Notably, we specify source and target compiler versions of 1.8,
instead of 1.7.

We also change from com.codahale.metrics to io.dropwizard.metrics,
since the former appears to be unmaintained, deprecated, and/or
archived (on github), and the latter is the new package, picking up
where the codahale package left off. We bump its version to 3.1.2
to match the metrics-core version that dropwizard 0.9.2 depends on.

When we bump dropwizard versions at some point in the future, we'll
have to come back here and bump metrics-core version again.